Tweaked buoyancy, made resources a tag instead of layer and made it function from trigger rather than collider

+++ Assets/Scripts/AmmoManager.cs
@@ -40,7 +40,7 @@ public class AmmoManager : MonoBehaviour {
         return (availableAmmo > 0);
     }
 
-    public int getAmmo()
+    public float getAmmo()
     {
         return availableAmmo;
     }
